+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+### 0.0.8
+- Tag pushing
+- Fixes bug with single key-stroke menu navigation
+
 ### 0.0.7
 - Adds show-refs command
 - Apply- and discard-at-point entities in show-ref view